+++ src/main/java/io/pismo/interview/creditcard/entity/Account.java
@@ -6,7 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.math.BigDecimal;
 
 @Entity
 @Table(name = "accounts")
@@ -25,6 +24,6 @@ public class Account {
     private String documentNumber;
 
     @Column(name = "available_credit_limit", nullable = false)
-    private BigDecimal availableCreditLimit;
+    private double availableCreditLimit;
 }
